Redesign the debug page (landscape)
