+++ composeApp/src/commonMain/composeResources/values/strings.xml
@@ -12,7 +12,7 @@
     <string name="run">Run</string>
     <string name="format">Format</string>
     <string name="reset">Reset</string>
-    <string name="reset_new_seed">Reset with new seed</string>
+    <string name="reset_new_seed">New seed</string>
     <string name="step_forward">Step forward</string>
     <string name="step_over">Step over</string>
     <string name="code">Code</string>
